+++ tests/test_get_auc.py
@@ -1,5 +1,6 @@
 import pandas as pd
 import pytest
+
 from pyplier.auc import get_auc
 from pyplier.plier_res import PLIERResults